cleaning up datamodel
