+++ DataModel/SuedtirolDBSpecial.cs
@@ -281,4 +281,6 @@ namespace DataModel
         public string Id { get; set; }
         public string Name { get; set; }
     }
+
+
 }
